+++ src/commands/build-umd.js
@@ -3,31 +3,8 @@ import path from 'path'
 
 import getUserConfig from '../getUserConfig'
 import {UserError} from '../errors'
+import {createBanner, createWebpackExternals} from '../utils'
 import webpackBuild from '../webpackBuild'
-
-function createBanner(pkg) {
-  let banner = `${pkg.name} ${pkg.version}`
-  if (pkg.homepage) {
-    banner += ` - ${pkg.homepage}`
-  }
-  if (pkg.license) {
-    banner += `\n${pkg.license} Licensed`
-  }
-  return banner
-}
-
-function createWebpackExternals(externals) {
-  return Object.keys(externals).reduce((webpackExternals, packageName) => {
-    let globalName = externals[packageName]
-    webpackExternals[packageName] = {
-      root: globalName,
-      commonjs2: packageName,
-      commonjs: packageName,
-      amd: packageName
-    }
-    return webpackExternals
-  }, {})
-}
 
 /**
  * Create a web module's UMD builds.